Updated how services are created to be more generic
