+++ Mittons.Fixtures/Attributes/IOptionAttribute.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Mittons.Fixtures.Models;
 
-namespace Mittons.Fixtures.Docker.Attributes
+namespace Mittons.Fixtures.Attributes
 {
     public interface IOptionAttribute
     {
