+++ Mittons.Fixtures/Containers/Attributes/SftpUserAccountAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Mittons.Fixtures.Docker.Attributes
+namespace Mittons.Fixtures.Containers.Attributes
 {
     [AttributeUsage(System.AttributeTargets.Property, AllowMultiple = true)]
     public class SftpUserAccountAttribute : Attribute
