+++ Mittons.Fixtures/Containers/IContainerService.cs
@@ -1,0 +1,7 @@
+namespace Mittons.Fixtures.Containers
+{
+    public interface IContainerService : IService
+    {
+        string ContainerId { get; }
+    }
+}
