--- Mittons.Fixtures/Docker/Gateways/NetworkGateway.cs
@@ -1,36 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Mittons.Fixtures.Extensions;
-using Mittons.Fixtures.Models;
-
-namespace Mittons.Fixtures.Docker.Gateways
-{
-    public class NetworkGateway : INetworkGateway
-    {
-        public async Task CreateAsync(string networkName, IEnumerable<Option> options, CancellationToken cancellationToken)
-        {
-            using (var process = new DockerProcess($"network create {options.ToExecutionParametersFormattedString()} {networkName}"))
-            {
-                await process.RunProcessAsync(cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        public async Task RemoveAsync(string networkName, CancellationToken cancellationToken)
-        {
-            using (var process = new DockerProcess($"network rm {networkName}"))
-            {
-                await process.RunProcessAsync(cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        public async Task ConnectAsync(string networkName, string containerId, string alias, CancellationToken cancellationToken)
-        {
-            using (var process = new DockerProcess($"network connect --alias {alias} {networkName} {containerId}"))
-            {
-                await process.RunProcessAsync(cancellationToken).ConfigureAwait(false);
-            }
-        }
-    }
-}
