+++ Mittons.Fixtures/IServiceGateway.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Mittons.Fixtures.Resources
+namespace Mittons.Fixtures
 {
     /// <summary>
     /// A gateway for managing <see cref="Mittons.Fixtures.Resources.IService">IServices</see>.
@@ -42,56 +42,5 @@ namespace Mittons.Fixtures.Resources
         /// This operation will remove the <see cref="Mittons.Fixtures.Resources.IService"/> and release all resources it held on the Host system.
         /// </remarks>
         Task RemoveServiceAsync(TService service, CancellationToken cancellationToken);
-    }
-
-    public interface IService : IAsyncLifetime
-    {
-        IEnumerable<Attribute> Attributes { get; }
-
-        IEnumerable<IServiceResource> ServiceResources { get; }
-    }
-
-    public interface IDockerService : IService
-    {
-        string Id { get; }
-
-        string ContainerId { get; }
-    }
-
-    /// <summary>
-    /// Represents details needed for the Host environment to communicate with the Guest <see cref="Mittons.Fixtures.Resources.IService"/>.
-    /// </summary>
-    /// <remarks>
-    /// These communications details can be a variety of mechanisms, such as file changes or network messaging.
-    /// </remarks>
-    public interface IServiceResource
-    {
-        /// <summary>
-        /// Gets the details for how the <see cref="Mittons.Fixtures.Resources.IService"/> monitors a resource to trigger actions.
-        /// </summary>
-        /// <remarks>
-        /// This can represent asynchronous communication mechanisms such as monitoring a file for changes or more synchronous request-respones channels such as http connections.
-        /// </remarks>
-        /// <returns>
-        /// Details describing the resource being monitored by the <see cref="Mittons.Fixtures.Resources.IServiceResource"/>.
-        /// </returns>
-        /// <value>
-        /// A <see href="https://en.wikipedia.org/wiki/Uniform_Resource_Identifier">Uniform Resource Identifier</see> with all known details on how the <see cref="Mittons.Fixtures.Resources.IServiceResource"/> listens for action triggers.
-        /// </value>
-        Uri GuestUri { get; }
-
-        /// <summary>
-        /// Gets the details for how the Host can communicate with an <see cref="Mittons.Fixtures.Resources.IService"/>.
-        /// </summary>
-        /// <remarks>
-        /// This can represent asynchronous communication mechanisms such as monitoring a file for changes or more synchronous request-respones channels such as http connections.
-        /// </remarks>
-        /// <returns>
-        /// Details describing the resource that can be accessed by the Host to communicate with the <see cref="Mittons.Fixtures.Resources.IService"/>.
-        /// </returns>
-        /// <value>
-        /// A <see href="https://en.wikipedia.org/wiki/Uniform_Resource_Identifier">Uniform Resource Identifier</see> with all known details on how the Host communicates with the <see cref="Mittons.Fixtures.Resources.IService"/>.
-        /// </value>
-        Uri HostUri { get; }
     }
 }
